Major Update
So i've created bookmark page, meal-planner page, made the icon of bookmark, calendar (meal planner actually but we used calendar icon), and sharing features. Also in this code, submit button in meal-planner should go to MyProfile which the page is held by Yosua. Also changed some of the pictures name.